+++ src/behaviours/Modal/TriggerModal.tsx
@@ -4,7 +4,6 @@ import ReactDOM from 'react-dom'
 import { Trigger } from '../../client/types'
 import { useMixpanel } from '../../context/MixpanelContext'
 import { useCollectorMutation } from '../../hooks/api/useCollectorMutation'
-import { useSeenMutation } from '../../hooks/api/useSeenMutation'
 import { useBrand } from '../../hooks/useBrandConfig'
 import { useCollector } from '../../hooks/useCollector'
 import { HandleCloseOptions } from './Modal.types'
@@ -27,18 +26,12 @@ const Modal = ({ trigger }: Props) => {
   const { mutate: collect } = useCollectorMutation()
 
   const brand = useBrand()
-  const { mutate: runSeen, isSuccess, isLoading } = useSeenMutation()
+  // useSeen({ trigger, skip: !open })
 
   useEffect(() => {
-    if (!open) return
-    if (invocationTimeStamp) return
-    if (isSuccess) return
-    if (isLoading) return
+    if (!!invocationTimeStamp) return
 
-    // seen gets called multiple times since Collector currently
-    // like to over-rerender componets. This timeout prevents from firing a ton
     const tId = setTimeout(() => {
-      runSeen(trigger)
       if (!invocationTimeStamp) {
         setInvocationTimeStamp(new Date().toISOString())
       }
@@ -47,7 +40,7 @@ const Modal = ({ trigger }: Props) => {
     return () => {
       clearTimeout(tId)
     }
-  }, [open, isSuccess, isLoading])
+  }, [invocationTimeStamp])
 
   if (!open) {
     return null
