+++ src/behaviours/Modal/modals/StandardModal.tsx
@@ -3,6 +3,7 @@ import React, { useEffect, useState } from 'react'
 import { isMobile } from 'react-device-detect'
 import CloseButton from '../../../components/CloseButton'
 import { useLogging } from '../../../context/LoggingContext'
+import { useSeen } from '../../../hooks/api/useSeenMutation'
 import { useBrandColors } from '../../../hooks/useBrandConfig'
 import {
   getIsModalFullyClickable,
@@ -30,6 +31,9 @@ const StandardModal = ({
   } = useModalDimensionsBasedOnImage({
     imageURL
   })
+  const isImageBrokenDontShowModal = !width || !height
+
+  useSeen({ trigger, skip: !stylesLoaded || isImageBrokenDontShowModal })
 
   const appendResponsiveBehaviour = React.useCallback(() => {
     return isMobile
@@ -263,7 +267,7 @@ const StandardModal = ({
     return null
   }
 
-  if (!width || !height) {
+  if (isImageBrokenDontShowModal) {
     error(
       "StandardModal: Couldn't get image dimensions, so not showing trigger. Investigate."
     )
