+++ src/visitors/bootstrap.tsx
@@ -100,6 +100,7 @@ export const bootstrapVisitor = ({
       visitorId = vidParam.split('?')[0]
     }
     visitor.id = visitorId
+    setCookie(cookieAccountJWT, visitorId || '', cookieValidDays)
 
     const sourceId = urlParams.get('source_id')
     if (sourceId) visitor.sourceId = sourceId
